incoming/review/pyproject.toml
[project]
name = "agent-review"
version = "0.1.0"
description = "Local Git diff review UI with inline comments stored under .agent-review and CLI export for agentic coders"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Author", email = "author@example.com" }]
dependencies = [
  "typer>=0.12.3",
  "rich>=13.9.0",
  "pyyaml>=6.0.0",
  "python-dotenv>=1.0.1",
  "fastapi>=0.115.0",
  "uvicorn>=0.32.0",
  "jinja2>=3.1.0",
  "pydantic>=2.6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=4.1.0",
  "pytest-mock>=3.12.0",
  "pytest-timeout>=2.4.0",
  "pytest-asyncio>=0.24.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0",
  "mypy>=1.11.0",
  "types-PyYAML>=6.0.0",
]

[project.scripts]
agent-review = "agent_review.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_review"]

[tool.hatch.build.targets.sdist]
include = ["src/agent_review"]

[tool.uv]
package = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
timeout = 30
addopts = ["-v", "--tb=short", "--strict-markers"]
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
  "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["src/agent_review"]
branch = true
omit = [
  "*/tests/*",
  "*/__pycache__/*",
  "src/agent_review/cli.py",
]

[tool.coverage.report]
fail_under = 75
show_missing = true
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "C4", "UP", "PT"]
ignore = ["E501", "B008", "B904"]

[tool.ruff.lint.per-file-ignores]
"src/agent_review/git.py" = ["S603", "S607"]

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
warn_redundant_casts = true
exclude = ["tests/"]
